Craigslist Deals Project with Kyle Coniker
This was the first working prototype of the program we made over the summer of 2016. We wanted to find deals on craigslist (specifically cars & trucks). It worked when our Twilio trials were still active, but has since become buggy. I put together the scraper to get the info, Kyle put together all the API interactions (with Edmunds and Twilio).